Unify HCAL navigator access between CPU and GPU using a new class PFHCALDenseIdNavigatorCore.
